+++ akshare/__init__.py
@@ -3151,9 +3151,13 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.17.42 fix: fix reits_realtime_em interface
 1.17.43 fix: fix fund_rating_all interface
 1.17.44 fix: fix stock_individual_spot_xq interface
+1.17.45 fix: fix stock_dxsyl_em interface
+1.17.46 fix: fix stock_financial_us_analysis_indicator_em interface
+1.17.47 fix: fix get_receipt interface
+1.17.48 fix: fix stock_hk_dividend_payout_em interface
 """
 
-__version__ = "1.17.44"
+__version__ = "1.17.48"
 __author__ = "AKFamily"
 
 import sys
@@ -3176,6 +3180,16 @@ if sys.version_info < (3, 9):
 del sys
 
 """
+中国外汇交易中心暨全国银行间同业拆借中心-基准-外汇市场-外汇掉期曲线-外汇掉漆 C-Swap 定盘曲线
+"""
+from akshare.fx.fx_c_swap_cm import fx_c_swap_cm
+
+"""
+上海证券交易所-产品-股票期权-信息披露-当日合约
+"""
+from akshare.option.option_current_sse import option_current_day_sse
+
+"""
 东方财富-A股-财务分析-主要指标
 """
 from akshare.stock_fundamental.stock_finance_sina import stock_financial_analysis_indicator_em
@@ -3189,6 +3203,11 @@ from akshare.option.option_margin import option_margin, option_margin_symbol
 东方财富-港股-证券资料
 """
 from akshare.stock.stock_profile_em import stock_hk_company_profile_em, stock_hk_security_profile_em
+
+"""
+东方财富-港股-核心必读
+"""
+from akshare.stock.stock_profile_em import stock_hk_dividend_payout_em, stock_hk_financial_indicator_em
 
 """
 东方财富网-行情中心-债券市场-质押式回购
@@ -3757,6 +3776,11 @@ from akshare.currency.currency_safe import currency_boc_safe
 from akshare.option.option_risk_indicator_sse import option_risk_indicator_sse
 
 """
+期权-上海证券交易所-当日合约
+"""
+from akshare.option.option_risk_indicator_sse import option_risk_indicator_sse
+"""
+
 全球宏观事件
 """
 from akshare.news.news_baidu import (
@@ -4766,7 +4790,6 @@ from akshare.stock_feature.stock_a_pe_and_pb import (
     stock_index_pe_lg,
 )
 from akshare.stock_feature.stock_a_indicator import (
-    stock_a_indicator_lg,
     stock_hk_indicator_eniu,
 )
 from akshare.stock_feature.stock_a_high_low import stock_a_high_low_statistics